feat: layer 전체 적용 기능 버튼 위치 변경
